Added enemy waves control track & game dialogue control track
added control track for some initial enemy waves. the timing is still not synchronized, this will be addressed in a future commit post prototyping.

added control track for general grizlok dialogue, and imported some assets for dialogue & image. these assets may be replaced in a future commit but will be used for initial design and prototyping.