adjusted ship rotation code
User Feedback: Aiming was a bit wonky
Based on user feedback, the ship's rotation function was adjusted to only couple the ship's rotation with the pitch & yaw (x & y axis inputs). The z axis rotation was set to 0 so that the only rotation in effect is the full rig rotation controlled via the Timeline animation. This made the aiming more controlled and addresses some of the feedback that aiming was a bit wonky